Ataque curto para os inimigos

+++ Assets/Scripts/Creatures/Attacker.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [RequireComponent(typeof(Creature))]
 public class Attacker : MonoBehaviour
 {
-    public int strengh = 0;
+    public float strengh = 0;
 
     public void Hit(Creature target)
     {
